+++ aioautomower/__init__.py
@@ -1,6 +1,5 @@
 """Automower library using aiohttp."""
 __all__ = [
-    "api",
     "auth",
     "cli",
     "const",
