+++ aioautomower/session.py
@@ -1,114 +1,93 @@
-"""Legacy Module to connect to Automower with websocket."""
+"""Module to connect to Automower with websocket."""
 import asyncio
-import json
+import contextlib
 import logging
 import time
 from typing import Literal
 
-from . import rest
-from .const import (
-    HUSQVARNA_URL,
-    MARGIN_TIME,
-    MIN_SLEEP_TIME,
-    WS_URL,
-)
-
-from .model import (HeadlightModes)
+from .auth import AbstractAuth
+from .const import MARGIN_TIME, MIN_SLEEP_TIME, REST_POLL_CYCLE
+from .exceptions import NoDataAvailableException
+from .model import HeadlightModes, MowerList
 
 _LOGGER = logging.getLogger(__name__)
 
-REST_POLL_CYCLE_LE = 86400
-WS_STATUS_UPDATE_CYLE = 840.0
-WS_TOLERANCE_TIME = 20.0
 
-class AutomowerSession:
-    """Session"""
+class AutomowerEndpoint:
+    """Endpoint URLs for the AutomowerConnect API."""
+
+    mowers = "mowers/"
+    "List data for all mowers linked to a user."
+
+    actions = "mowers/{mower_id}/actions"
+    "Accepts actions to control a mower linked to a user."
+
+    calendar = "mowers/{mower_id}/calendar"
+    "Update the calendar on the mower."
+
+    settings = "mowers/{mower_id}/settings"
+    "Update the settings on the mower."
+
+    stay_out_zones = "mowers/{mower_id}/stayOutZones/{stay_out_id}"
+    "Enabe or disable the stay-out zone."
+
+    work_area_calendar = "mowers/{mower_id}/workAreas{work_area_id}/calendar"
+    "Update the calendar for a work area on the mower."
+
+
+class AutomowerApi:
+    """Automower API to communicate with an Automower.
+
+    The `AutomowerApi` is the primary API service for this library. It supports
+    operations like getting a status or sending commands.
+    """
 
     def __init__(
         self,
-        api_key: str,
-        token: dict = None,
-        low_energy=True,
-        ws_heartbeat_interval: float = 60.0,
-        loop=None,
+        auth: AbstractAuth,
+        poll: bool = False,
     ) -> None:
         """Create a session.
 
-        :param str api_key: A 36 digit api key.
-        :param dict token: A token as returned by rest.GetAccessToken.async_get_access_token()
-        :param float ws_heartbeat_interval: Periodicity of keep-alive pings on the websocket in seconds.
-        :param loop: Event-loop for task execution. If None, the event loop in the current OS thread is used.
+        :param class auth: The AbstractAuth class from aioautomower.auth.
+        :param bool poll: Poll data with rest if True.
         """
-        _LOGGER.warning("The AutomowerSession class is depracated. Please migrate to AutomowerApi class")
-        self.api_key = api_key
-        self.token = token
+        self.auth = auth
+        self.poll = poll
         self.data_update_cbs = []
         self.token_update_cbs = []
-        self.ws_heartbeat_interval = ws_heartbeat_interval
         self.rest_task = False
-        self.low_energy = low_energy
-        if loop is None:
-            self.loop = asyncio.get_event_loop()
-        else:
-            self.loop = loop
-
-        self.data = None
-
+        self.loop = asyncio.get_event_loop()
+        self.token = None
+        self.data = {}
+        self.mowers = {}
+        self.listen_task = None
         self.ws_task = None
+
         self.token_task = None
-        self.websocket_monitor_task = None
         self.rest_task = None
 
-    def register_data_callback(self, callback, schedule_immediately=False):
-        """Register a data update callback.
-
-        :param func callback: Callback fired on data updates. Takes one dict argument which is the up-to-date mower data list.
-        :param bool schedule_immediately: Schedule callback immediately (if data is available).
-        """
+    def register_data_callback(self, callback):
+        """Register a data update callback."""
         if callback not in self.data_update_cbs:
             self.data_update_cbs.append(callback)
-        if schedule_immediately:
-            self._schedule_data_callback(
-                callback, delay=1e-3
-            )  # Need a delay for home assistant to finish entity setup.
+
+    def _schedule_data_callback(self, cb, delay=0.0):
+        if self.poll and self.data is None:
+            raise NoDataAvailableException
+        self.loop.call_later(delay, cb, self.mowers)
+
+    def _schedule_data_callbacks(self):
+        for cb in self.data_update_cbs:
+            self._schedule_data_callback(cb)
 
     def unregister_data_callback(self, callback):
         """Unregister a data update callback.
 
-        :param func callback: Callback fired on data updates. Takes one dict argument which is the up-to-date mower data list.
+        :param func callback: Takes one function, which should be unregistered.
         """
         if callback in self.data_update_cbs:
             self.data_update_cbs.remove(callback)
-
-    def register_token_callback(self, callback, schedule_immediately=False):
-        """Register a token update callback.
-
-        :param func callback: Callback fired on token updates. Takes one dict argument which is the newly received token.
-        :param bool schedule_immediately: Schedule callback immediately (if token is available).
-        """
-        if callback not in self.token_update_cbs:
-            self.token_update_cbs.append(callback)
-        if schedule_immediately:
-            self._schedule_token_callback(
-                callback, delay=1e-3
-            )  # Need a delay for home assistant to finish entity setup.
-
-    async def logincc(self, client_secret: str) -> dict:
-        """Login with client credentials.
-
-        This method gets an access token with a client_id (Api key) and a client_secret.
-        This token can't be refreshed. Create a new one after it is expired.
-
-        :param str client_secret: Your client_secret
-        :return dict: The token as returned by
-        rest.GetAccessTokenClientCredentials.async_get_access_token().
-        You can store this persistently and pass it to the constructor
-        on subsequent instantiations.
-        """
-        a = rest.GetAccessTokenClientCredentials(self.api_key, client_secret)
-        self.token = await a.async_get_access_token()
-        self._schedule_token_callbacks()
-        return self.token
 
     async def connect(self):
         """Connect to the API.
@@ -119,51 +98,20 @@ class AutomowerSession:
         token is created with the Authorization Code Grant. Call this method
         before any other methods.
         """
-        if self.token is None:
-            raise AttributeError("No token to connect with.")
-        if time.time() > (self.token["expires_at"] - MARGIN_TIME):
-            await self.refresh_token()
-
-        self.data = await self.get_status()
         self._schedule_data_callbacks()
 
-        if "amc:api" not in self.token["scope"]:
-            _LOGGER.error(
-                "Your API-Key is not compatible to the websocket, please refresh it on %s",
-                HUSQVARNA_URL,
-            )
-        else:
-            self.ws_task = self.loop.create_task(self._ws_task())
-        self.rest_task = self.loop.create_task(self._rest_task())
-        self.token_task = self.loop.create_task(self._token_monitor_task())
+        if self.poll:
+            await self.get_status()
+            self.rest_task = self.loop.create_task(self._rest_task())
 
-    async def ws_and_token_session(self):
-        """Connect to the API.
-
-        This method handles the login and starts a task that keep the access
-        token constantly fresh. This method works only, if the token is created with the
-        Authorization Code Grant. Call this method before any other methods.
-        """
-        if self.token is None:
-            raise AttributeError("No token to connect with.")
-        if time.time() > (self.token["expires_at"] - MARGIN_TIME):
-            await self.refresh_token()
-
-        self.data = await self.get_status()
-        self._schedule_data_callbacks()
-
-        if "amc:api" not in self.token["scope"]:
-            raise NotImplementedError()
-        else:
-            self.ws_task = self.loop.create_task(self._ws_task())
-        self.token_task = self.loop.create_task(self._token_monitor_task())
+        self.ws_task = self.loop.create_task(self.auth.websocket())
+        self.auth.register_ws_callback(self.callback)
 
     async def close(self):
         """Close the session."""
         for task in [
             self.ws_task,
             self.token_task,
-            self.websocket_monitor_task,
             self.rest_task,
         ]:
             tasks = []
@@ -171,101 +119,85 @@ class AutomowerSession:
                 tasks.append(task)
                 if not task.cancelled():
                     task.cancel()
-        try:
+        with contextlib.suppress(asyncio.CancelledError):
             await asyncio.gather(*tasks)
-        except asyncio.CancelledError:
-            pass
 
-    async def get_status(self):
+    async def get_status(self) -> MowerList:
         """Get mower status via Rest."""
-        if self.token is None:
-            _LOGGER.warning("No token available")
-            return None
-        d = rest.GetMowerData(
-            self.api_key,
-            self.token["access_token"],
-            self.token["provider"],
-            self.token["token_type"],
-        )
-        return await d.async_mower_state()
-
-    async def action(self, mower_id: str, payload: str, command_type: str):
-        """Send command to the mower via Rest."""
-        if self.token is None:
-            _LOGGER.warning("No token available")
-            return None
-        a = rest.Return(
-            self.api_key,
-            self.token["access_token"],
-            self.token["provider"],
-            self.token["token_type"],
-            mower_id,
-            payload,
-            command_type,
-        )
-        return await a.async_mower_command()
+        mower_list = await self.auth.get_json(AutomowerEndpoint.mowers)
+        for idx, _ent in enumerate(mower_list["data"]):
+            mower_list["data"][idx]["attributes"].update(
+                mower_list["data"][idx]["attributes"]["settings"]
+            )
+            del mower_list["data"][idx]["attributes"]["settings"]
+        self.data = mower_list
+        self.mower_as_dict_dataclass()
+        return self.mowers
 
     async def resume_schedule(self, mower_id: str):
-        """Removes any ovveride on the Planner and let the mower
-        resume to the schedule set by the Calendar"""
-        command_type = "actions"
-        payload = {"data": {"type": "ResumeSchedule"}}
-        try:
-            await self.send_command_via_rest(mower_id, payload, command_type)
-        except rest.CommandNotPossibleError as exception:
-            _LOGGER.error("Command couldn't be sent to the command que: %s", exception)
+        """Resume schedule.
+
+        Remove any ovveride on the Planner and let the mower
+        resume to the schedule set by the Calendar.
+        """
+        body = {"data": {"type": "ResumeSchedule"}}
+        url = AutomowerEndpoint.actions.format(mower_id=mower_id)
+        await self.auth.post_json(url, json=body)
 
     async def pause_mowing(self, mower_id: str):
         """Send pause mowing command to the mower via Rest."""
-        command_type = "actions"
-        payload = {"data": {"type": "Pause"}}
-        await self.send_command_via_rest(mower_id, payload, command_type)
+        body = {"data": {"type": "Pause"}}
+        url = AutomowerEndpoint.actions.format(mower_id=mower_id)
+        await self.auth.post_json(url, json=body)
 
     async def park_until_next_schedule(self, mower_id: str):
         """Send park until next schedule command to the mower."""
-        command_type = "actions"
-        payload = {"data": {"type": "ParkUntilNextSchedule"}}
-        await self.send_command_via_rest(mower_id, payload, command_type)
+        body = {"data": {"type": "ParkUntilNextSchedule"}}
+        url = AutomowerEndpoint.actions.format(mower_id=mower_id)
+        await self.auth.post_json(url, json=body)
 
     async def park_until_further_notice(self, mower_id: str):
         """Send park until further notice command to the mower."""
-        command_type = "actions"
-        payload = {"data": {"type": "ParkUntilFurtherNotice"}}
-        await self.send_command_via_rest(mower_id, payload, command_type)
+        body = {"data": {"type": "ParkUntilFurtherNotice"}}
+        url = AutomowerEndpoint.actions.format(mower_id=mower_id)
+        await self.auth.post_json(url, json=body)
 
     async def park_for(self, mower_id: str, duration_in_min: int):
-        """Parks the mower for a period of minutes. The mower will drive to
-        the charching station and park for the duration set by the command."""
-        command_type = "actions"
-        payload = {
+        """Parks the mower for a period of minutes.
+
+        The mower will drive to
+        the charching station and park for the duration set by the command.
+        """
+        body = {
             "data": {
                 "type": "Park",
                 "attributes": {"duration": duration_in_min},
             }
         }
-        await self.send_command_via_rest(mower_id, payload, command_type)
+        url = AutomowerEndpoint.actions.format(mower_id=mower_id)
+        await self.auth.post_json(url, json=body)
 
     async def start_for(self, mower_id: str, duration_in_min: int):
         """Start the mower for a period of minutes."""
-        command_type = "actions"
-        payload = {
+        body = {
             "data": {
                 "type": "Park",
                 "attributes": {"duration": duration_in_min},
             }
         }
-        await self.send_command_via_rest(mower_id, payload, command_type)
+        url = AutomowerEndpoint.actions.format(mower_id=mower_id)
+        await self.auth.post_json(url, json=body)
 
     async def set_cutting_height(self, mower_id: str, cutting_height: int):
         """Start the mower for a period of minutes."""
-        command_type = "settings"
-        payload = {
+        body = {
             "data": {
                 "type": "settings",
                 "attributes": {"cuttingHeight": cutting_height},
             }
         }
-        await self.send_command_via_rest(mower_id, payload, command_type)
+        url = AutomowerEndpoint.settings.format(mower_id=mower_id)
+        await self.auth.post_json(url, json=body)
 
     async def set_headlight_mode(
         self,
@@ -278,14 +210,14 @@ class AutomowerSession:
         ],
     ):
         """Send headlight mode to the mower."""
-        command_type = "settings"
-        payload = {
+        body = {
             "data": {
                 "type": "settings",
                 "attributes": {"headlight": {"mode": headlight_mode}},
             }
         }
-        await self.send_command_via_rest(mower_id, payload, command_type)
+        url = AutomowerEndpoint.settings.format(mower_id=mower_id)
+        await self.auth.post_json(url, json=body)
 
     async def set_calendar(
         self,
@@ -293,51 +225,30 @@ class AutomowerSession:
         task_list: list,
     ):
         """Send calendar task to the mower."""
-        command_type = "calendar"
-        payload = {
+        body = {
             "data": {
                 "type": "calendar",
                 "attributes": {"tasks": task_list},
             }
         }
-        await self.send_command_via_rest(mower_id, payload, command_type)
+        url = AutomowerEndpoint.calendar.format(mower_id=mower_id)
+        await self.auth.post_json(url, json=body)
 
-    async def send_command_via_rest(
-        self, mower_id: str, payload: dict, command_type: str
+    async def switch_stay_out_zone(
+        self, mower_id: str, stay_out_zone_id: str, switch: bool
     ):
-        """Send a command to the mower."""
-        json_payload = json.dumps(payload)
-        rest_init = rest.Return(
-            self.api_key,
-            self.token["access_token"],
-            self.token["provider"],
-            self.token["token_type"],
-            mower_id,
-            json_payload,
-            command_type,
+        """Enable or disable a stay out zone."""
+        body = {
+            "data": {
+                "type": "stayOutZone",
+                "id": stay_out_zone_id,
+                "attributes": {"enable": switch},
+            }
+        }
+        url = AutomowerEndpoint.stay_out_zones.format(
+            mower_id=mower_id, stay_out_id=stay_out_zone_id
         )
-        try:
-            await rest_init.async_mower_command()
-        except rest.CommandNotPossibleError as exception:
-            _LOGGER.error("Command couldn't be sent to the command que: %s", exception)
-
-    async def invalidate_token(self):
-        """Invalidate token via Rest."""
-        if self.token is None:
-            _LOGGER.warning("No token available")
-            return None
-        token = rest.RevokeAccessToken(self.token["access_token"])
-        return await token.async_delete_access_token()
-
-    async def refresh_token(self):
-        """Refresh token via Rest."""
-        if "refresh_token" not in self.token:
-            _LOGGER.warning("No refresh token available")
-            return None
-        _LOGGER.debug("Refresh access token")
-        r = rest.RefreshAccessToken(self.api_key, self.token["refresh_token"])
-        self.token = await r.async_refresh_access_token()
-        self._schedule_token_callbacks()
+        await self.auth.patch_json(url, json=body)
 
     async def _token_monitor_task(self):
         while True:
@@ -349,46 +260,51 @@ class AutomowerSession:
 
             _LOGGER.debug("token_monitor_task sleeping for %s sec", sleep_time)
             await asyncio.sleep(sleep_time)
-            await self.refresh_token()
+            await self.auth.async_get_access_token()
 
     def _update_data(self, j):
         if self.data is None:
-            _LOGGER.error("Failed to update data with ws response (no data)")
-            return
-        for datum in self.data["data"]:
-            if datum["type"] == "mower" and datum["id"] == j["id"]:
-                if j["type"] == "positions-event":
-                    last_pos_identical = (
-                        datum["attributes"]["positions"][0]
-                        == j["attributes"]["positions"][0]
-                    )
-                    if not last_pos_identical:
-                        j["attributes"]["positions"].extend(
-                            datum["attributes"]["positions"]
+            raise NoDataAvailableException
+        if self.data is not None:
+            for datum in self.data["data"]:
+                if datum["type"] == "mower" and datum["id"] == j["id"]:
+                    if j["type"] == "positions-event":
+                        last_pos_identical = (
+                            datum["attributes"]["positions"][0]
+                            == j["attributes"]["positions"][0]
                         )
-                        _LOGGER.debug(
-                            "j['attributes']['positions']: %s",
-                            j["attributes"]["positions"],
-                        )
-                for attrib in j["attributes"]:
-                    try:
-                        tasks = j["attributes"]["calendar"]["tasks"]
-                        if len(tasks) == 0:
-                            temp_task = datum["attributes"]["calendar"]["tasks"]
+                        if not last_pos_identical:
+                            j["attributes"]["positions"].extend(
+                                datum["attributes"]["positions"]
+                            )
+                    for attrib in j["attributes"]:
+                        try:
+                            tasks = j["attributes"]["calendar"]["tasks"]
+                            if len(tasks) == 0:
+                                temp_task = datum["attributes"]["calendar"]["tasks"]
+                                datum["attributes"][attrib] = j["attributes"][attrib]
+                                datum["attributes"]["calendar"]["tasks"] = temp_task
+                            if len(tasks) > 0:
+                                datum["attributes"][attrib] = j["attributes"][attrib]
+                        except KeyError:
                             datum["attributes"][attrib] = j["attributes"][attrib]
-                            datum["attributes"]["calendar"]["tasks"] = temp_task
-                        if len(tasks) > 0:
-                            datum["attributes"][attrib] = j["attributes"][attrib]
-                    except KeyError:
-                        datum["attributes"][attrib] = j["attributes"][attrib]
-                return
-        _LOGGER.error("Failed to update data with ws response (id not found)")
+        self.mower_as_dict_dataclass()
+        self._schedule_data_callbacks()
 
-    def _schedule_token_callback(self, cb, delay=0.0):
-        if self.token is None:
-            _LOGGER.debug("No token available. Will not schedule callback")
-            return
-        self.loop.call_later(delay, cb, self.token)
+    def mower_as_dict_dataclass(self):
+        """Convert mower data to a dictionary DataClass."""
+        mowers_list = MowerList(**self.data)
+        for mower in mowers_list.data:
+            self.mowers[mower.id] = mower.attributes
 
-    def _schedule_token_callbacks(self):
-        for cb in self.token_update_cbs:
+    def callback(self, new_data):
+        """Pass received websocket data to the update function."""
+        if new_data:
+            self._update_data(new_data)
+
+    async def _rest_task(self):
+        """Poll data periodically via Rest."""
+        while True:
+            await asyncio.sleep(REST_POLL_CYCLE)
+            self.data = await self.get_status()
+            self._schedule_data_callbacks()
